Hi all,
  please review this change to avoid CodeCache triggered GCs temporarily being ignored.

This is a backport of https://github.com/openjdk/jdk/pull/26189, which did not apply cleanly.

The issue has been the hunk in `gcCause.hpp`, and in `G1Policy::record_concurrent_mark_remark_end()` where the context was wrong.

Testing: tier1-5, with the follow-up patch to the test case

Thanks,
  Thomas